requirements.txt
python-dotenv
langchain_google_genai
langchain_community
ipykernel
bs4
arxivloader
pymupdf
arxiv
wikipedia
langchain-text-splitters
chromadb
notebook
jupyter
langchain_huggingface
sentence-transformers
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_huggingface
duckdb
pandas
validators
unstructured
numexpr
huggingface_hub
fastapi
langserve
sse_starlette
pydantic
uvicorn
tenacity
passlib[bcrypt]
python-jose
transformers
pypdf2
phidata
yfinance
packaging
groq
pgvector
psycopg[binary]
requests==2.32.3
typing-extensions==4.12.2
setuptools==70.0.0
starlette>=0.40.0
langgraph-prebuilt==0.2.2
